backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
openai
websockets==12.0
cryptography
python-dotenv
aiofiles
ijson
jsonpatch